Finally got this to compile on Mac OS X with C++11 flags.

+++ src/SFML/Window/OSX/SFWindowController.mm
@@ -458,7 +458,7 @@
     if (*m_fullscreenMode != sf::VideoMode()) {
         CGLContextObj cgcontext = (CGLContextObj)[context CGLContextObj];
         
-        GLint dim[2] = {m_fullscreenMode->width, m_fullscreenMode->height};
+        GLint dim[2] = {static_cast<int>(m_fullscreenMode->width), static_cast<int>(m_fullscreenMode->height)};
         
         CGLSetParameter(cgcontext, kCGLCPSurfaceBackingSize, dim);
         CGLEnable(cgcontext, kCGLCESurfaceBackingSize);
